Updated GRID implementation in compiler
Update the grammar to have not put the grid params into the general
lexer params. The problem is still that we the channel reads all the
component parameters for the grid.

In addition to that I fixed the issue where I wasn’t querying the wrong
index for the components